Enhance create_star_ref.py to support gzip input for fasta files, add sequence name filtering, and improve biotype filtering with detailed status reporting.

+++ scripts/create_star_ref.py
@@ -1,8 +1,10 @@
 #!/usr/bin/env python3
 import os
 import sys
+import gzip
 import argparse
 from typing import Tuple, List, Dict, Set
+from collections import Counter
 
 
 class CustomFormatter(argparse.ArgumentDefaultsHelpFormatter, argparse.RawDescriptionHelpFormatter):
@@ -148,10 +150,8 @@ def parse_args() -> argparse.Namespace:
     # example
     ./scripts/create_star_ref.py \
       --organism "Macaca mulatta" \
-      --fasta /home/nickyoungblut/tmp/genomes/Macaca_mulatta.Mmul_10.dna.toplevel.fa.gz \
+      --fasta /home/nickyoungblut/tmp/genomes/reference_sources/Macaca_mulatta.Mmul_10.dna.toplevel.fa \
       /home/nickyoungblut/tmp/genomes/reference_sources/Macaca_mulatta.Mmul_10.113.gtf
-      
-      
     """
     parser = argparse.ArgumentParser(description=desc, epilog=epi, formatter_class=CustomFormatter)
     parser.add_argument(
@@ -175,11 +175,23 @@ def parse_args() -> argparse.Namespace:
     )
     return parser.parse_args()
 
-def process_gtf_line(line: str, outF: 'TextIO', biotypes: Set[str], exclude_tags: List[str], status: Dict[str, int]):
+def process_gtf_line(
+        line: str, 
+        outF: 'TextIO', 
+        biotypes: Set[str], 
+        exclude_tags: List[str], 
+        seq_names: Set[str], 
+        status: Dict[str, int]
+    ):
     """
     Process a single gtf line.
     Args:
-        line
+        line: GTF line.
+        outF: Output file handle.
+        biotypes: Set of biotypes to keep.
+        exclude_tags: List of tags to exclude.
+        seq_names: Set of sequence names to keep.
+        status: Dictionary to store status.
     """
     # simply write out header line
     if line.startswith("#"):
@@ -202,6 +214,9 @@ def process_gtf_line(line: str, outF: 'TextIO', biotypes: Set[str], exclude_tags
                 continue
     fields = fields[:8]
 
+    # store seq names
+    seq_names.add(fields[0])
+
     # convert gene_id
     if attributes.get("gene_id") and not attributes.get("gene_version"):
         try:
@@ -212,17 +227,55 @@ def process_gtf_line(line: str, outF: 'TextIO', biotypes: Set[str], exclude_tags
             pass
     
     # filter by biotype
-    if attributes.get("gene_type") and attributes.get("gene_type") not in biotypes:
-        status["biotype"] += 1
-        return None
-    elif attributes.get("transcript_type") and attributes.get("transcript_type") not in biotypes:
-        status["biotype"] += 1
-        return None
+    biotype_fields = ['gene_biotype', 'gene_type', 'transcript_type', 'transcript_biotype']
+    for field in biotype_fields:
+        if attributes.get(field) and attributes[field] not in biotypes:
+            status["biotype"] += 1
+            try:
+                status[attributes[field]] += 1
+            except KeyError:
+                status[attributes[field]] = 1
+            return None
+    
     # filter by tags
     if attributes.get("tag") and attributes.get("tag") in exclude_tags:
         status["tag"] += 1
         return None
 
+    # write out
+    attributes = "; ".join([f"{k} \"{v}\"" for k, v in attributes.items()])
+    outF.write("\t".join(fields + [attributes]) + "\n")
+
+def process_fasta(fasta: str, output_fasta: str, seq_names: Set[str]):
+    """
+    Process a fasta file. Check to make sure that the sequence names are in the set.
+    Args:
+        fasta: Path to input fasta file.
+        output_fasta: Path to output fasta file.
+        seq_names: Set of sequence names to keep.
+    """
+    print(f"Processing fasta: {os.path.basename(fasta)}", file=sys.stderr)
+
+    if fasta.endswith(".gz"):
+        _open = gzip.open
+    else:
+        _open = open
+
+    with _open(fasta) as inF, gzip.open(output_fasta, 'wb') as outF:
+        write = False
+        for line in inF:
+            try:
+                line = line.decode()   
+            except AttributeError:
+                pass
+            if line.startswith(">"):
+                if line.lstrip(">").split(" ")[0] in seq_names:
+                    write = True
+                else:
+                    write = False
+                    print(f"Sequence not in GTF: {line.strip()}", file=sys.stderr)
+            if write:
+                outF.write(line.encode())
 
 def main():
     # parse cli arguments
@@ -233,7 +286,15 @@ def main():
     if not os.path.exists(args.output_dir):
         os.makedirs(args.output_dir, exist_ok=True)
 
+    # check input
+    if args.fasta and not os.path.exists(args.fasta):
+        sys.exit(f"Error: {args.fasta} not found")
+    if not os.path.exists(args.gtf):
+        sys.exit(f"Error: {args.gtf} not found")
+
     # iterate over gtf
+    print(f"Processing GTF: {os.path.basename(args.gtf)}", file=sys.stderr)
+    seq_names = set()
     status = {"total_raw": 0, "biotype": 0, "tag": 0}
     output_gtf = os.path.join(args.output_dir, f"{args.organism}.gtf")
     with open(args.gtf) as inF, open(output_gtf, 'w') as outF:
@@ -242,17 +303,25 @@ def main():
                 line, outF, 
                 biotypes=biotype_index[args.organism],
                 exclude_tags=args.exclude_tags,
+                seq_names=seq_names,
                 status=status,
             )
         
-    # status
+    ## GTF processing status
     print(f"Total records in GTF: {status['total_raw']}", file=sys.stderr)
     for key in ["biotype", "tag"]:
         print(f"Filtered {status[key]} records by {key}", file=sys.stderr)
+    print("-- Count of biotypes filtered --", file=sys.stderr)
+    for key in status:
+        if key not in ["total_raw", "biotype", "tag"]:
+            print(f"{key}: {status[key]}", file=sys.stderr)
+
+    # fasta
+    if args.fasta:
+        output_fasta = os.path.join(args.output_dir, f"{args.organism}.fna.gz")
+        process_fasta(args.fasta, output_fasta, seq_names)
 
 
 # main
 if __name__ == "__main__":
-    #from dotenv import load_dotenv
-    #load_dotenv(override=True)
     main()
